requirements.txt
wheel==0.38.4
dataclasses==0.6
greenlet==2.0.1
grpcio==1.49.1
grpcio-tools==1.49.1
protobuf==4.21.12
psycopg2-binary==2.9.5
pydantic==1.10.2
six==1.16.0
SQLAlchemy==1.4.44
typing_extensions==4.4.0
